requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
werkzeug>=2.3.0

# PDF Processing
PyPDF2==3.0.1
reportlab==4.0.7
pdfplumber==0.10.3
pytesseract==0.3.10

# Core API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Storage
minio>=7.2.0
boto3>=1.34.0

# Image Processing (for future modules)
Pillow>=10.1.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# Monitoring
prometheus-client>=0.19.0